[refactor] #84 common 패키지로 예외 핸들러 및 코드 이동
