+++ src/main/java/com/nuclear/boomm/product/domain/ExtraDescription.java
@@ -1,7 +1,7 @@
 package com.nuclear.boomm.product.domain;
 
 import com.nuclear.boomm.product.error.CustomException;
-import com.nuclear.boomm.product.error.ErrorCode;
+import com.nuclear.boomm.common.error.ErrorCode;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
