+++ src/main/java/com/nuclear/boomm/product/error/CustomException.java
@@ -1,5 +1,6 @@
 package com.nuclear.boomm.product.error;
 
+import com.nuclear.boomm.common.error.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
